Attempting to fix config persistence issue.
Problem was that we were trying to update the config before it had been read into memory. if you do
this, config is None.
Instead, we now do a check to make sure that a user has actually altered an input field before
trying to do the update.

This issue may affect other dcc.Store objects...